requirements.txt
redis>=5.0
PyYAML>=6.0
